+++ packages/apptive_grid_core/lib/network/filter/filter_expressions.dart
@@ -25,7 +25,8 @@ class Today extends _FilterExpression with ComparableFilterableMixin {
 
 /// Filter columns by the logged in user
 /// This works for [DataType.createdBy] and [DataType.assignee]
-class LoggedInUser extends _FilterExpression {
+/// If filtering for a field with [DataType.createdBy] you should combine it with [ActorFilter]
+class LoggedInUser extends _FilterExpression with ActorFilterableMixin {
   /// Creates a Filter Expression to check the loggedIn User
   const LoggedInUser() : super(expression: 'loggedInUser()');
 }